move the generic BLE layer to internal/shared/ble
The reusable half of internal/cli/ble was never CLI-specific, and one
consumer already proved it: shared/discovery imports it to identify Wendy
Lite boards, which is a shared/ package reaching up into cli/. Moving the
generic packages down removes that edge and makes the layer available to
anything, not just the CLI.

  shared/ble/central/  the Connection API (GATT + L2CAP) and the
                       L2CAP-to-net.Conn adapter TLS runs over
  shared/ble/scan/     the continuous scanner
  cli/ble/             the Wendy protocol clients, and only those

Only the generic code moves. agent_client.go, agent_tls_over_ble.go and
lite_client.go stay in cli/ble: they are Wendy protocol and the CLI is
their only consumer.

lite_info.go moves even though it is Wendy-specific, because leaving it
behind would keep the shared/-to-cli/ edge this commit exists to remove —
shared/discovery reads a board's L2CAP PSM through it. It sits at the
shared/ble root rather than in central/ or scan/, so those two stay free
of Wendy identifiers, and its package doc says why it is there.

Both directories keep `package ble`. That is legal and invisible here:
nothing in cli/ble referenced lite_info.go's symbols, and no consumer
uses both halves — helpers.go and wifi.go take only the agent and Lite
clients, while liteclient, shared/discovery and providers take only the
info service. So no call site changes beyond an import path. Both package
docs note that a future file needing both must alias.

No behaviour change. Content edits are confined to import paths and to
doc comments naming a path that moved; scan_darwin.h's reference to
ble_darwin.m was already stale before this.

Verified: go build, vet, gofmt and tests clean, plus the GOOS=windows
cross-compile gate.
