+++ go/internal/shared/ble/central/ble_darwin.go
