+++ go/internal/shared/ble/scan/scan_darwin.go
