+++ go/internal/shared/ble/scan/scan_linux.go
