+++ go/internal/shared/ble/scan/scan_test.go
