+++ go/internal/shared/ble/scan/scan_windows.go
