+++ go/internal/shared/ble/scan/uuid.go
