+++ go/internal/shared/ble/scan/uuid_test.go
