+++ go/internal/shared/ble/scan/winrt_parse_test.go
